fix(kotlin): simplify kotlin codes
